+++ src/operators/query/element/exists.ts
@@ -9,15 +9,14 @@ export const $exists = (selector: string, value: Any, _options: Options) => {
   const b = !!value;
   // top-level keys and array elements.
   if (!nested || selector.match(/\.\d+$/)) {
-    const opts = { pathArray: selector.split(".") };
-    return (o: AnyObject) => (resolve(o, selector, opts) !== undefined) === b;
+    return (o: AnyObject) => (resolve(o, selector) !== undefined) === b;
   }
   // for nested keys we resolve the entire value path so we don't confuse array scalars with plural values.
   const parentSelector = selector.substring(0, selector.lastIndexOf("."));
-  const opts = { pathArray: parentSelector.split("."), preserveIndex: true };
+  const opts = { preserveIndex: true };
   return (o: AnyObject) => {
     const path = resolveGraph(o, selector, opts) as AnyObject;
-    const val = resolve(path, parentSelector, opts);
+    const val = resolve(path, parentSelector);
     return isArray(val)
       ? val.some(v => v !== undefined) === b
       : (val !== undefined) === b;